Replacing explicit target member with "target"

+++ ZFramework/ZWinPaletteDialog.cpp
@@ -462,8 +462,8 @@ ZWinPaletteDialog* ZWinPaletteDialog::ShowPaletteDialog(std::string sCaption, ZG
     pDialog->mBehavior |= ZWinDialog::eBehavior::Draggable | ZWinDialog::eBehavior::OKButton | ZWinDialog::eBehavior::CancelButton;
     pDialog->mStyle = gDefaultDialogStyle;
     pDialog->mStyle.fp = gStyleButton.fp;
-    pDialog->mStyle.paddingV = gSpacer;
-    pDialog->mStyle.paddingH = gSpacer;
+    pDialog->mStyle.paddingV = (int32_t)gSpacer;
+    pDialog->mStyle.paddingH = (int32_t)gSpacer;
 
     pDialog->mpColorMap = pColorMap;
     pDialog->mOriginalColorMap = *pColorMap;
